chore: abstract to utils module
