+++ package.json
@@ -5,16 +5,19 @@
     "packages": []
   },
   "dependencies": {
-    "@bazel/ibazel": "0.15.10",
     "react": "17.0.2",
     "react-dom": "17.0.2"
   },
   "devDependencies": {
     "@bazel/bazelisk": "1.10.1",
+    "@bazel/ibazel": "0.15.10",
+    "@bazel/typescript": "4.4.6",
+    "@types/node": "16.11.11",
     "esbuild-loader": "2.16.0",
     "husky": "7.0.4",
     "lint-staged": "12.1.2",
     "prettier": "2.5.0",
+    "typescript": "4.5.2",
     "webpack": "5.60.0",
     "webpack-cli": "4.9.1"
   },
